Remove mention of poles. This was an error.
A late-night thinko cause me to code up some of the polylog
continuations incorrectly: I'd snuck in a complex conjugate,
without realizing. This made zeros look like poles (because
the phase was reversed.) Magnitude showed they were zeros,
not poles and that's when I realized my mistake. So it goes.